+++ app/assets/javascripts/application.js
@@ -16,3 +16,9 @@
 //= require popper
 //= require bootstrap
 //= require_tree .
+
+$(document).ready(function(){
+	$(function () { //Enables Bootstrap Tooltips Everywhere
+	$('[data-toggle="tooltip"]').tooltip()
+	});
+});