+++ app/assets/javascripts/custom.js
@@ -1,0 +1,5 @@
+$(document).on("turbolinks:load",function(){
+  $(function () { //Enables Bootstrap Tooltips Everywhere
+	  $('[data-toggle="tooltip"]').tooltip()
+	});
+})