+++ app/controllers/api/v1/users/passwords_controller.rb
@@ -1,0 +1,6 @@
+module Api::V1
+	class Users::PasswordsController < DeviseTokenAuth::PasswordsController
+		include DeviseTokenAuth::Concerns::SetUserByToken
+		protect_from_forgery with: :null_session
+	end
+end