+++ app/controllers/main_site_base_controller.rb
@@ -1,0 +1,2 @@
+class MainSiteBaseController < ApplicationController
+end