+++ app/controllers/test_sessions_controller.rb
@@ -1,4 +1,4 @@
-class TestSessionsController < ApplicationController
+class TestSessionsController < MainSiteBaseController
 	def new
 	end
 end